Added removing and adding, tested in main activity (remove later)
