+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\hunte\.android\avd\Pixel_2_API_30.avd" />
+            <value value="C:\Users\hunte\.android\avd\Pixel_4_API_33.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-10-26T19:07:55.946761400Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-10-27T00:11:43.821254500Z" />
   </component>
 </project>
